+++ examples/unix/c11/z_pub_thr.c
@@ -25,7 +25,7 @@ int main(int argc, char **argv) {
         exit(-1);
     }
     char *keyexpr = "test/thr";
-    size_t len = atoi(argv[1]);
+    size_t len = (size_t)atoi(argv[1]);
     uint8_t *value = (uint8_t *)malloc(len);
     memset(value, 1, len);
